Add Function to Retrieve API Data
Add function 'get_data' which accepts the parameters
'url' and 'headers'. These parameters are used to
access the endpoint for the API from Stockanalysis.
The function retrieves the data from the API, and
parses it to JSON, and returns the data to be readable
for other aspects of the program.