Add Basic Logic for Email
Add the basic logic for the user entering
their email address. Takes the input entered
by the user to check if it was null, and then
pressed the confirm button. If the user's input
was null and they pressed the confirm button,
a popup displays stating that they did not
enter an email address. If user enters a email
that is not null and presses confirm button,
a popup displays stating that it is working
(will change soon)