Add Email Function and Update Logic
Add a function called 'check_email' that accepts
the input entered from the GUI as a variable
and checks it against a regular expression to
determine if the user entered a valid email format
or not. This function returns true for a valid email
format and false for an invalid email format. Update
the logic in loop to run GUI to account for this email
validation function.